+++ board/safety/safety_hyundai_common.h
@@ -1,12 +1,12 @@
 #ifndef SAFETY_HYUNDAI_COMMON_H
 #define SAFETY_HYUNDAI_COMMON_H
 
-const int HYUNDAI_PARAM_EV_GAS = 1;
-const int HYUNDAI_PARAM_HYBRID_GAS = 2;
-const int HYUNDAI_PARAM_LONGITUDINAL = 4;
-const int HYUNDAI_PARAM_CAMERA_SCC = 8;
-const int HYUNDAI_PARAM_CANFD_HDA2 = 16;
-const int HYUNDAI_PARAM_ALT_LIMITS = 64; // TODO: shift this down with the rest of the common flags
+const uint8_t HYUNDAI_PARAM_EV_GAS = 1;
+const uint8_t HYUNDAI_PARAM_HYBRID_GAS = 2;
+const uint8_t HYUNDAI_PARAM_LONGITUDINAL = 4;
+const uint8_t HYUNDAI_PARAM_CAMERA_SCC = 8;
+const uint8_t HYUNDAI_PARAM_CANFD_HDA2 = 16;
+const uint8_t HYUNDAI_PARAM_ALT_LIMITS = 64; // TODO: shift this down with the rest of the common flags
 
 const uint8_t HYUNDAI_PREV_BUTTON_SAMPLES = 8;  // roughly 160 ms
 const uint32_t HYUNDAI_STANDSTILL_THRSLD = 12;  // 0.375 kph
